backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic-settings==2.10.1
httpx==0.28.1
openai==1.101.0
supabase==2.18.1
python-jose[cryptography]==3.5.0
slowapi==0.1.9
redis==6.4.0
pyyaml==6.0.2
orjson==3.11.2
